Fix some casts invalid on 64-bit windows.

+++ src/Platform.cpp
@@ -37,7 +37,7 @@ namespace platform {
 		if (hasConsole()) return false;
 #ifdef WIN32
 		int hConHandle;
-		long lStdHandle;
+		intptr_t lStdHandle;
 		CONSOLE_SCREEN_BUFFER_INFO coninfo;
 		FILE *fp;
 		// allocate a console for this app
@@ -47,19 +47,19 @@ namespace platform {
 		coninfo.dwSize.Y = 500;
 		SetConsoleScreenBufferSize(GetStdHandle(STD_OUTPUT_HANDLE), coninfo.dwSize);
 		// redirect unbuffered STDOUT to the console
-		lStdHandle = (long)GetStdHandle(STD_OUTPUT_HANDLE);
+		lStdHandle = (intptr_t)GetStdHandle(STD_OUTPUT_HANDLE);
 		hConHandle = _open_osfhandle(lStdHandle, _O_TEXT);
 		fp = _fdopen( hConHandle, "w" );
 		*stdout = *fp;
 		setvbuf( stdout, NULL, _IONBF, 0 );
 		// redirect unbuffered STDIN to the console
-		lStdHandle = (long)GetStdHandle(STD_INPUT_HANDLE);
+		lStdHandle = (intptr_t)GetStdHandle(STD_INPUT_HANDLE);
 		hConHandle = _open_osfhandle(lStdHandle, _O_TEXT);
 		fp = _fdopen( hConHandle, "r" );
 		*stdin = *fp;
 		setvbuf( stdin, NULL, _IONBF, 0 );
 		// redirect unbuffered STDERR to the console
-		lStdHandle = (long)GetStdHandle(STD_ERROR_HANDLE);
+		lStdHandle = (intptr_t)GetStdHandle(STD_ERROR_HANDLE);
 		hConHandle = _open_osfhandle(lStdHandle, _O_TEXT);
 		fp = _fdopen( hConHandle, "w" );
 		*stderr = *fp;
